Refactor location and file names
