+++ sample-saml-query-profile/saml-query-profile-client/sample-request-messages/assertion-id-request.xml
