+++ sample-saml-query-profile/saml-query-profile-client/sample-request-messages/attribute-query-request.xml
