+++ sample-saml-query-profile/saml-query-profile-client/sample-request-messages/authnquery-session-response.xml
