+++ sample-saml-query-profile/saml-query-profile-client/sample-request-messages/authnquery-with-session-index-response.xml
