+++ sample-saml-query-profile/saml-query-profile-client/sample-request-messages/dummy-attribute-query.xml
