+++ sample-saml-query-profile/saml-query-profile-client/sample-request-messages/subject-query.xml
